requirements.txt
apache-airflow==2.8.1
apache-airflow-providers-docker==3.9.0
pyspark==3.5.1
kafka-python==2.0.2
websocket-client==1.7.0 # Finnhub Producer

# Visualization
streamlit==1.32.0
pandas==1.5.3
pyarrow==15.0.0